requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
base58==2.1.1
binance==0.3
binance-connector==1.13.0
bitarray==2.5.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
constantly==15.1.0
cryptography==37.0.2
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
h11==0.12.0
hexbytes==0.2.2
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.0
outcome==1.2.0
pandas==1.3.5
parsimonious==0.8.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rfc3986==1.5.0
rlp==2.0.1
selenium==4.2.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toolz==0.11.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
web3==5.25.0
webdriver-manager==3.7.0
websockets==9.1
wsproto==1.1.0
yarl==1.7.2
zope.interface==5.4.0
